Cadastro de Eventos - Ajuste ao carregar  campo tipo eventos
