+++ src/SME.SGP.WebClient/src/paginas/CalendarioEscolar/Eventos/cadastro/campos/tipoEventoCadastroEventos.js
@@ -1,17 +1,15 @@
 import PropTypes from 'prop-types';
-import React, { useCallback, useContext, useEffect } from 'react';
+import React, { useContext, useEffect } from 'react';
 import { SelectComponent } from '~/componentes';
 import { OPCAO_TODOS } from '~/constantes';
 import ListaLocalOcorrencia from '~/constantes/localOcorrencia';
 import tipoEvento from '~/dtos/tipoEvento';
-import { api } from '~/servicos';
 import EventosCadastroContext from '../eventosCadastroContext';
 
 const TipoEventoCadastroEventos = ({ form, onChangeCampos, desabilitar }) => {
   const {
     listaTipoEvento,
     setListaTipoEvento,
-    setListaTipoEventoOrigem,
     setDesabilitarLetivo,
     listaTipoEventoOrigem,
   } = useContext(EventosCadastroContext);
@@ -19,24 +17,6 @@ const TipoEventoCadastroEventos = ({ form, onChangeCampos, desabilitar }) => {
   const { dreId, ueId, tipoEventoId } = form.values;
 
   const nomeCampo = 'tipoEventoId';
-
-  const obterTiposEvento = useCallback(async () => {
-    const tiposEvento = await api.get(
-      'v1/calendarios/eventos/tipos/listar?ehCadastro=true&numeroRegistros=100'
-    );
-    if (tiposEvento?.data?.items) {
-      setListaTipoEvento([...tiposEvento.data.items]);
-      setListaTipoEventoOrigem([...tiposEvento.data.items]);
-    } else {
-      setListaTipoEvento([]);
-      setListaTipoEventoOrigem([]);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  useEffect(() => {
-    obterTiposEvento();
-  }, [obterTiposEvento]);
 
   useEffect(() => {
     if (listaTipoEvento?.length === 1 && ueId) {
